+++ NEWS.md
@@ -60,7 +60,7 @@ phylopath 1.0.1
 phylopath 1.0.0
 --------------------------------------------------------------------------------
 
-* Citation info now points to the [bioRxiv paper](https://www.biorxiv.org/content/early/2017/10/31/212068).
+* Citation info now points to the [bioRxiv paper](https://www.biorxiv.org/content/10.1101/212068v1).
 
 * All modeling functions now completely rely on the `phylolm` package, and no 
   longer use `ape`. This is a major change, that will possibly change the