+++ NEWS.md
@@ -6,6 +6,8 @@ phylopath 1.1.3.9000
 * Fixed a bug where the range of the width scale for paths in `plot.fitted_DAG` 
   was incorrectly set to the `max(weight)`, instead of `max(abs(weight))`. 
   (Thanks Yu Xu for the report.)
+  
+* Better user messaging and documentation around the use of the `boot` parameter.
 
 phylopath 1.1.3
 --------------------------------------------------------------------------------